+++ src/functions/model/make-generator-quicktype-input-data.ts
@@ -1,11 +1,7 @@
-import type { WorkspaceContext } from '@causa/workspace';
+import { callDeferred, type WorkspaceContext } from '@causa/workspace';
 import type { InputData } from 'quicktype-core';
-import { makeJsonSchemaInputData } from '../../code-generation/index.js';
 import type { ModelConfiguration } from '../../configurations/index.js';
-import {
-  ModelMakeGeneratorQuicktypeInputData,
-  ModelParseCodeGeneratorInputs,
-} from '../../definitions/index.js';
+import { ModelMakeGeneratorQuicktypeInputData } from '../../definitions/index.js';
 
 /**
  * Implements the {@link ModelMakeGeneratorQuicktypeInputData} function for JSONSchema models.
@@ -14,15 +10,7 @@ import {
  */
 export class ModelMakeGeneratorQuicktypeInputDataForJsonSchema extends ModelMakeGeneratorQuicktypeInputData {
   async _call(context: WorkspaceContext): Promise<InputData> {
-    const { files, nestedSchemas, includeFullReferences } = await context.call(
-      ModelParseCodeGeneratorInputs,
-      { configuration: this.configuration },
-    );
-
-    return await makeJsonSchemaInputData(files, {
-      nestedSchemas,
-      includeFullReferences,
-    });
+    return await callDeferred(this, context, import.meta.url);
   }
 
   _supports(context: WorkspaceContext): boolean {