+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+Breaking changes:
+
+- Move backfill, code generation, and services as separate exports.
+
+Features:
+
+- Export `makeJsonSchemaInputDataFromSources`.
+
+Chores:
+
+- Defer loading of heavy dependencies (`quicktype-core`, `@scalar/*`) during function registration.
+
 ## v0.26.1 (2026-02-06)
 
 Chore: